Fixed RecommendCombos Mutating Global toonOrgs
RecommendCombos class was mutating the global toonOrgs object (not intended behavior). This would change the ordering of the global toonOrgs object, causing toons' organic gags to be incorrectly stored and updated.

Now, RecommendCombos makes a copy of the global toonOrgs object to use and mutate, preserving the state of the global object.