Custom Scrollbar Class, Mobile Layout Fix
Mobile layout '.mobile' component is set at 100vh, and set to display flex. '.mobile .page-wrap' is set to height: 100% so that it scales to fill all available space. '#mobile-nav' is no longer set to a fixed position, since it is accounted for in the flex container.

Added '.custom-scollbar' CSS class, since it is a reused style.