Fixed Saved Toon's Current Organic Gag Button Highlight
Previously, the OrganicPicker component would always initialize activeBtn as 0, meaning that a saved toon's active button would always be 'None' until you updated it to another gag.

Now, pass in the toon's organic gag as a parameter, and set activeBtn based off of the org gag's index, or default to 0.