Added Cog Persistance
When the user initially sets the cog level, a random suit and name are chosen, as before.

Now, cogSlice stores these cog suit and cog name values, and the cog component builds the cog object with those 3 pieces of data, instead of just level as before.

Tweaked the cog class to accept suit and name parameters, defaulting to random.