Restructured Gags Data
Instead of gagsData being of the form gags[organic][track] (which required gags objects to be copy/pasted twice), gagsData is now just gags[track] and the relevant stat for each gag is a dictionary with an organic and non-organic key to their corresponding values.

Reworked components that relied on the previous data structure.