Toons Component Bugfix
Used to, if you reset the toons while a Toon Panel's Organic Picker was active, the state of that Toon Panel would not reset, causing its content to be inaccessible until you refreshed the page.

Now, added a useReducer to manage each Toon Panel's Organic Picker active status in the parent ToonsComponent. When the reset button is clicked, it resets all Organic Pickers active status back to false before resetting the toons state.