Add Local Storage Clear if Outdated
To prevent any bugs caused by major redux refactoring while site is still in development, set a last refactored key in local storage and check if it matches a "last updated" value. If it doesn't, clear the local storage state object. If it does, keep the local storage state object.