Fix FindCombo Output Inconsistencies
Previously, a number of bugs occurred in the UI output of gags' Lured and Combo multipliers. This was due to the algorithm cloning the combo object, and the way in which the algorithm would mutate the clone and merge it back in.

This clone would still reference the original, and changes to the new combo would also not be reflected in the final output. Interestingly, in this case, it wouldn't affect the combos' calculations for accuracy and damage, but would result in some gags displaying correct information in the UI.

Solution was to not dynamically mutate the class instances, but instead to store the data that needed to be mutated outside of the class instance, then create a new instance with the updated data in each iteration.